Refactor folder structure and put each component in a folder
Corrections on new props

Test fail on pipeline

mend
