+++ .gitignore
@@ -1,2 +1,3 @@
 .terraform
 .idea
+.gradle
