+++ packages/frontend/app/src/session/Sessions/sessionCard/Session.test.tsx
@@ -22,6 +22,7 @@ beforeEach(() => {
       time={time}
       onEditClicked={jest.fn()}
       isEditing={false}
+      forceUpdate={jest.fn()}
     />
   );
 });
