+++ packages/frontend/app/src/session/SessionsContainer.css
@@ -3,5 +3,5 @@
 }
 
 .ui.modal > .content {
-  background-color: gold;
+  background-color: grey;
 }
